Rename tests to test_dg_* format.
