+++ tests/test_dg_terms_calls.py
@@ -17,7 +17,7 @@ from sfepy.discrete.common.dof_info import EquationMap
 from sfepy.terms.terms_dg import AdvectionDGFluxTerm, \
     NonlinearHyperbolicDGFluxTerm, NonlinearScalarDotGradTerm, \
     DiffusionDGFluxTerm, DiffusionInteriorPenaltyTerm
-from test_field_dg import prepare_dgfield_1D, prepare_field_2D
+from test_dg_field import prepare_dgfield_1D, prepare_field_2D
 
 
 class Test(TestCommon):
